Add new player 'on-boarding' logic
After new player signs up, the 'show_new_player' view is displayed.
He has the ability to challenge anyone on the ladder 2 times.
If he wins a challenge match, he takes the plase of the defeated player.
If he loses both challenge matches, he is placed at the bottom of the
ladder.